+++ workshops/introduction-to-event-sourcing/exercises/src/test/java/io/eventdriven/introductiontoeventsourcing/e04_getting_state_from_events/esdb/mutable/GettingStateFromEventsTests.java
@@ -11,6 +11,7 @@ import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 
 import static io.eventdriven.introductiontoeventsourcing.e04_getting_state_from_events.esdb.mutable.GettingStateFromEventsTests.ShoppingCartEvent.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class GettingStateFromEventsTests extends EventStoreDBTest {
@@ -133,8 +134,8 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
       this.status = status;
     }
 
-    public List<PricedProductItem> productItems() {
-      return productItems;
+    public PricedProductItem[] productItems() {
+      return productItems.toArray(PricedProductItem[]::new);
     }
 
     public void setProductItems(List<PricedProductItem> productItems) {
@@ -198,14 +199,9 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
 
     assertEquals(shoppingCartId, shoppingCart.id());
     assertEquals(clientId, shoppingCart.clientId());
-    assertEquals(2, shoppingCart.productItems().size());
+    assertEquals(2, shoppingCart.productItems().length);
 
-    assertEquals(shoesId, shoppingCart.productItems().get(0).productId());
-    assertEquals(pairOfShoes.quantity(), shoppingCart.productItems().get(0).quantity());
-    assertEquals(pairOfShoes.unitPrice(), shoppingCart.productItems().get(0).unitPrice());
-
-    assertEquals(tShirtId, shoppingCart.productItems().get(1).productId());
-    assertEquals(tShirt.quantity(), shoppingCart.productItems().get(1).quantity());
-    assertEquals(tShirt.unitPrice(), shoppingCart.productItems().get(1).unitPrice());
+    assertThat(shoppingCart.productItems()[0]).usingRecursiveComparison().isEqualTo(pairOfShoes);
+    assertThat(shoppingCart.productItems()[1]).usingRecursiveComparison().isEqualTo(tShirt);
   }
 }
