+++ workshops/introduction-to-event-sourcing/exercises/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/ProductItems/ProductPriceCalculator.java
@@ -1,0 +1,7 @@
+package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.ProductItems;
+
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.ProductItems.ProductItems.*;
+
+public interface ProductPriceCalculator {
+  PricedProductItem calculate(ProductItem productItems);
+}
