+++ workshops/introduction-to-event-sourcing/exercises/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/ShoppingCartEvent.java
@@ -1,0 +1,37 @@
+package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable;
+
+import java.time.OffsetDateTime;
+import java.util.UUID;
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.ProductItems.ProductItems.*;
+
+public sealed interface ShoppingCartEvent {
+  record ShoppingCartOpened(
+    UUID shoppingCartId,
+    UUID clientId
+  ) implements ShoppingCartEvent {
+  }
+
+  record ProductItemAddedToShoppingCart(
+    UUID shoppingCartId,
+    PricedProductItem productItem
+  ) implements ShoppingCartEvent {
+  }
+
+  record ProductItemRemovedFromShoppingCart(
+    UUID shoppingCartId,
+    PricedProductItem productItem
+  ) implements ShoppingCartEvent {
+  }
+
+  record ShoppingCartConfirmed(
+    UUID shoppingCartId,
+    OffsetDateTime confirmedAt
+  ) implements ShoppingCartEvent {
+  }
+
+  record ShoppingCartCanceled(
+    UUID shoppingCartId,
+    OffsetDateTime canceledAt
+  ) implements ShoppingCartEvent {
+  }
+}
