+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e02_getting_state_from_events/mutable/GettingStateFromEventsTests.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.UUID;
 
 import static io.eventdriven.introductiontoeventsourcing.e02_getting_state_from_events.mutable.GettingStateFromEventsTests.ShoppingCartEvent.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class GettingStateFromEventsTests {
@@ -136,8 +137,8 @@ public class GettingStateFromEventsTests {
       this.status = status;
     }
 
-    public List<PricedProductItem> productItems() {
-      return productItems;
+    public PricedProductItem[] productItems() {
+      return productItems.toArray(PricedProductItem[]::new);
     }
 
     public void setProductItems(List<PricedProductItem> productItems) {
@@ -184,7 +185,7 @@ public class GettingStateFromEventsTests {
       var productId = pricedProductItem.productId();
       var quantityToAdd = pricedProductItem.quantity();
 
-      productItems().stream()
+      productItems.stream()
         .filter(pi -> pi.productId().equals(productId))
         .findAny()
         .ifPresentOrElse(
@@ -198,7 +199,7 @@ public class GettingStateFromEventsTests {
       var productId = pricedProductItem.productId();
       var quantityToRemove = pricedProductItem.quantity();
 
-      productItems().stream()
+      productItems.stream()
         .filter(pi -> pi.productId().equals(productId))
         .findAny()
         .ifPresentOrElse(
@@ -259,14 +260,9 @@ public class GettingStateFromEventsTests {
 
     assertEquals(shoppingCartId, shoppingCart.id());
     assertEquals(clientId, shoppingCart.clientId());
-    assertEquals(2, shoppingCart.productItems().size());
+    assertEquals(2, shoppingCart.productItems().length);
 
-    assertEquals(shoesId, shoppingCart.productItems().getFirst().productId());
-    assertEquals(pairOfShoes.quantity(), shoppingCart.productItems().get(0).quantity());
-    assertEquals(pairOfShoes.unitPrice(), shoppingCart.productItems().get(0).unitPrice());
-
-    assertEquals(tShirtId, shoppingCart.productItems().get(1).productId());
-    assertEquals(tShirt.quantity(), shoppingCart.productItems().get(1).quantity());
-    assertEquals(tShirt.unitPrice(), shoppingCart.productItems().get(1).unitPrice());
+    assertThat(shoppingCart.productItems()[0]).usingRecursiveComparison().isEqualTo(pairOfShoes);
+    assertThat(shoppingCart.productItems()[1]).usingRecursiveComparison().isEqualTo(tShirt);
   }
 }
