+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e04_getting_state_from_events/esdb/mutable/GettingStateFromEventsTests.java
@@ -14,6 +14,7 @@ import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 
 import static io.eventdriven.introductiontoeventsourcing.e04_getting_state_from_events.esdb.mutable.GettingStateFromEventsTests.ShoppingCartEvent.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class GettingStateFromEventsTests extends EventStoreDBTest {
@@ -120,54 +121,6 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
     public ShoppingCart() {
     }
 
-    public UUID id() {
-      return id;
-    }
-
-    public void setId(UUID id) {
-      this.id = id;
-    }
-
-    public UUID clientId() {
-      return clientId;
-    }
-
-    public void setClientId(UUID clientId) {
-      this.clientId = clientId;
-    }
-
-    public ShoppingCartStatus status() {
-      return status;
-    }
-
-    public void setStatus(ShoppingCartStatus status) {
-      this.status = status;
-    }
-
-    public List<PricedProductItem> productItems() {
-      return productItems;
-    }
-
-    public void setProductItems(List<PricedProductItem> productItems) {
-      this.productItems = productItems;
-    }
-
-    public OffsetDateTime confirmedAt() {
-      return confirmedAt;
-    }
-
-    public void setConfirmedAt(OffsetDateTime confirmedAt) {
-      this.confirmedAt = confirmedAt;
-    }
-
-    public OffsetDateTime canceledAt() {
-      return canceledAt;
-    }
-
-    public void setCanceledAt(OffsetDateTime canceledAt) {
-      this.canceledAt = canceledAt;
-    }
-
     public void evolve(ShoppingCartEvent event) {
       switch (event) {
         case ShoppingCartOpened opened -> apply(opened);
@@ -192,7 +145,7 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
       var productId = pricedProductItem.productId();
       var quantityToAdd = pricedProductItem.quantity();
 
-      productItems().stream()
+      productItems.stream()
         .filter(pi -> pi.productId().equals(productId))
         .findAny()
         .ifPresentOrElse(
@@ -206,7 +159,7 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
       var productId = pricedProductItem.productId();
       var quantityToRemove = pricedProductItem.quantity();
 
-      productItems().stream()
+      productItems.stream()
         .filter(pi -> pi.productId().equals(productId))
         .findAny()
         .ifPresentOrElse(
@@ -223,6 +176,54 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
     private void apply(ShoppingCartCanceled event) {
       setStatus(ShoppingCartStatus.Canceled);
       setConfirmedAt(event.canceledAt());
+    }
+
+    public UUID id() {
+      return id;
+    }
+
+    public void setId(UUID id) {
+      this.id = id;
+    }
+
+    public UUID clientId() {
+      return clientId;
+    }
+
+    public void setClientId(UUID clientId) {
+      this.clientId = clientId;
+    }
+
+    public ShoppingCartStatus status() {
+      return status;
+    }
+
+    public void setStatus(ShoppingCartStatus status) {
+      this.status = status;
+    }
+
+    public PricedProductItem[] productItems() {
+      return productItems.toArray(PricedProductItem[]::new);
+    }
+
+    public void setProductItems(List<PricedProductItem> productItems) {
+      this.productItems = productItems;
+    }
+
+    public OffsetDateTime confirmedAt() {
+      return confirmedAt;
+    }
+
+    public void setConfirmedAt(OffsetDateTime confirmedAt) {
+      this.confirmedAt = confirmedAt;
+    }
+
+    public OffsetDateTime canceledAt() {
+      return canceledAt;
+    }
+
+    public void setCanceledAt(OffsetDateTime canceledAt) {
+      this.canceledAt = canceledAt;
     }
   }
 
@@ -292,14 +293,9 @@ public class GettingStateFromEventsTests extends EventStoreDBTest {
 
     assertEquals(shoppingCartId, shoppingCart.id());
     assertEquals(clientId, shoppingCart.clientId());
-    assertEquals(2, shoppingCart.productItems().size());
+    assertEquals(2, shoppingCart.productItems().length);
 
-    assertEquals(shoesId, shoppingCart.productItems().getFirst().productId());
-    assertEquals(pairOfShoes.quantity(), shoppingCart.productItems().get(0).quantity());
-    assertEquals(pairOfShoes.unitPrice(), shoppingCart.productItems().get(0).unitPrice());
-
-    assertEquals(tShirtId, shoppingCart.productItems().get(1).productId());
-    assertEquals(tShirt.quantity(), shoppingCart.productItems().get(1).quantity());
-    assertEquals(tShirt.unitPrice(), shoppingCart.productItems().get(1).unitPrice());
+    assertThat(shoppingCart.productItems()[0]).usingRecursiveComparison().isEqualTo(pairOfShoes);
+    assertThat(shoppingCart.productItems()[1]).usingRecursiveComparison().isEqualTo(tShirt);
   }
 }
