+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/solution1/ProductItems/ProductItems.java
@@ -1,0 +1,92 @@
+package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution1.ProductItems;
+
+import java.util.UUID;
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+public final class ProductItems {
+  public static class ProductItem {
+    private UUID productId;
+    private int quantity;
+
+    public ProductItem(UUID productId, int quantity) {
+      this.setProductId(productId);
+      this.setQuantity(quantity);
+    }
+
+    public ProductItem(){}
+
+    public UUID productId() {
+      return productId;
+    }
+
+    public void setProductId(UUID productId) {
+      this.productId = productId;
+    }
+
+    public int quantity() {
+      return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+      this.quantity = quantity;
+    }
+
+    public void add(int quantity) {
+      this.quantity += quantity;
+    }
+
+    public void subtract(int quantity) {
+      this.quantity -= quantity;
+    }
+  }
+
+  public static class PricedProductItem {
+    private UUID productId;
+    private double unitPrice;
+    private int quantity;
+
+    public PricedProductItem(UUID productId, int quantity, double unitPrice) {
+      this.setProductId(productId);
+      this.setUnitPrice(unitPrice);
+      this.setQuantity(quantity);
+    }
+
+    public PricedProductItem(){}
+
+    private double totalAmount() {
+      return quantity() * unitPrice();
+    }
+
+    public UUID productId() {
+      return productId;
+    }
+
+    public void setProductId(UUID productId) {
+      this.productId = productId;
+    }
+
+    public double unitPrice() {
+      return unitPrice;
+    }
+
+    public void setUnitPrice(double unitPrice) {
+      this.unitPrice = unitPrice;
+    }
+
+    public int quantity() {
+      return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+      this.quantity = quantity;
+    }
+
+    public void add(int quantity) {
+      this.quantity += quantity;
+    }
+
+    public void subtract(int quantity) {
+      this.quantity -= quantity;
+    }
+  }
+}
