+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/solution2/Aggregate.java
@@ -1,0 +1,9 @@
+package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2;
+
+import java.util.UUID;
+
+public interface Aggregate<ShoppingCartEvent> {
+  UUID id();
+
+  void evolve(ShoppingCartEvent event);
+}
