+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/solution2/BusinessLogicTests.java
@@ -1,77 +1,24 @@
 package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2;
 
-import io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.BusinessLogic.FakeProductPriceCalculator;
-import io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.BusinessLogic.ShoppingCart;
-import io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.BusinessLogic.ShoppingCartStatus;
+import io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems.FakeProductPriceCalculator;
+import io.eventdriven.introductiontoeventsourcing.e05_business_logic.tools.EventStore;
 import org.junit.jupiter.api.Test;
 
-import java.time.OffsetDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.UUID;
 
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems.ProductItems.PricedProductItem;
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems.ProductItems.ProductItem;
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ShoppingCartEvent.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class BusinessLogicTests {
-  public sealed interface ShoppingCartEvent {
-    record ShoppingCartOpened(
-      UUID shoppingCartId,
-      UUID clientId
-    ) implements ShoppingCartEvent {
-    }
-
-    record ProductItemAddedToShoppingCart(
-      UUID shoppingCartId,
-      PricedProductItem productItem
-    ) implements ShoppingCartEvent {
-    }
-
-    record ProductItemRemovedFromShoppingCart(
-      UUID shoppingCartId,
-      PricedProductItem productItem
-    ) implements ShoppingCartEvent {
-    }
-
-    record ShoppingCartConfirmed(
-      UUID shoppingCartId,
-      OffsetDateTime confirmedAt
-    ) implements ShoppingCartEvent {
-    }
-
-    record ShoppingCartCanceled(
-      UUID shoppingCartId,
-      OffsetDateTime canceledAt
-    ) implements ShoppingCartEvent {
-    }
-  }
-
-  public record PricedProductItem(
-    UUID productId,
-    int quantity,
-    double unitPrice
-  ) {
-    public double totalAmount() {
-      return quantity * unitPrice;
-    }
-  }
-
-  public record ProductItem(
-    UUID productId,
-    int quantity
-  ) {
-  }
-
-  static ShoppingCart getShoppingCart(Object[] events) {
+  static ShoppingCart getShoppingCart(EventStore eventStore, UUID shoppingCartId) {
     // 1. Add logic here
-    var shoppingCartEvents = Arrays.stream(events)
-      .filter(ShoppingCartEvent.class::isInstance)
-      .map(ShoppingCartEvent.class::cast)
-      .toList();
+    var shoppingCart = ShoppingCart.initial();
 
-    var shoppingCart = new ShoppingCart();
-
-    for (var event : shoppingCartEvents) {
+    for (var event : eventStore.readStream(ShoppingCartEvent.class, shoppingCartId)) {
       shoppingCart.evolve(event);
     }
 
@@ -79,12 +26,11 @@ public class BusinessLogicTests {
   }
 
   @Test
-  public void gettingState_ForSequenceOfEvents_ShouldSucceed() {
+  public void runningSequenceOfBusinessLogic_ShouldGenerateSequenceOfEvents() {
     var shoppingCartId = UUID.randomUUID();
     var clientId = UUID.randomUUID();
     var shoesId = UUID.randomUUID();
     var tShirtId = UUID.randomUUID();
-
     var twoPairsOfShoes = new ProductItem(shoesId, 2);
     var pairOfShoes = new ProductItem(shoesId, 1);
     var tShirt = new ProductItem(tShirtId, 1);
@@ -95,65 +41,70 @@ public class BusinessLogicTests {
     var pricedPairOfShoes = new PricedProductItem(shoesId, 1, shoesPrice);
     var pricedTShirt = new PricedProductItem(tShirtId, 1, tShirtPrice);
 
-    var events = new ArrayList<>();
+    var eventStore = new EventStore();
 
     // Open
-    var result = ShoppingCart.open(shoppingCartId, clientId);
-    var shoppingCart = result.getValue();
-    events.add(result.getKey());
+    var opened = ShoppingCart.open(shoppingCartId, clientId);
+    eventStore.appendToStream(shoppingCartId, new Object[]{opened.getKey()});
 
-    // Add two pairs of shoes
-    shoppingCart = getShoppingCart(events.toArray());
-    events.add(
-      shoppingCart.addProduct(FakeProductPriceCalculator.returning(shoesPrice), twoPairsOfShoes)
+    // Add Two Pair of Shoes
+    var shoppingCart = getShoppingCart(eventStore, shoppingCartId);
+    ShoppingCartEvent result = shoppingCart.addProduct(
+      FakeProductPriceCalculator.returning(shoesPrice),
+      twoPairsOfShoes
     );
+    eventStore.appendToStream(shoppingCartId, new Object[]{result});
 
     // Add T-Shirt
-    shoppingCart = getShoppingCart(events.toArray());
-    events.add(
-      shoppingCart.addProduct(FakeProductPriceCalculator.returning(tShirtPrice), tShirt)
+    shoppingCart = getShoppingCart(eventStore, shoppingCartId);
+    result = shoppingCart.addProduct(
+      FakeProductPriceCalculator.returning(tShirtPrice),
+      tShirt
     );
+    eventStore.appendToStream(shoppingCartId, new Object[]{result});
 
-    // Remove pair of shoes
-    // Hack alert!
-    //
-    // See that's why immutability is so cool, as it's predictable
-    // As we're sharing objects (e.g. in PricedProductItem in events)
-    // then adding them into list and changing it while appending/removing
-    // then we can have unpleasant surprises.
-    //
-    // This will not likely happen if all objects are recreated (e.g. in the web requests)
-    // However when it happens then it's tricky to diagnose.
-    // Uncomment lines below and debug to find more.
-    // shoppingCart = getShoppingCart(events.toArray());
-    events.add(
-      shoppingCart.removeProduct(pricedPairOfShoes)
-    );
+    // Remove a pair of shoes
+    shoppingCart = getShoppingCart(eventStore, shoppingCartId);
+    result = shoppingCart.removeProduct(pricedPairOfShoes);
+    eventStore.appendToStream(shoppingCartId, new Object[]{result});
 
     // Confirm
-    // Uncomment line below and debug to find bug.
-    // shoppingCart = getShoppingCart(events.toArray());
-    events.add(
-      shoppingCart.confirm()
-    );
+    shoppingCart = getShoppingCart(eventStore, shoppingCartId);
+    result = shoppingCart.confirm();
+    eventStore.appendToStream(shoppingCartId, new Object[]{result});
 
-    // Cancel
-    ShoppingCart finalShoppingCart = shoppingCart;
-    assertThrows(IllegalStateException.class, () ->
-      events.add(
-        finalShoppingCart.cancel()
-      )
-    );
+    // Try Cancel
+    ShoppingCart finalShoppingCart = getShoppingCart(eventStore, shoppingCartId);
+    assertThrows(IllegalStateException.class, () -> {
+      var cancelResult = finalShoppingCart.cancel();
+      eventStore.appendToStream(shoppingCartId, new Object[]{cancelResult});
+    });
 
-    // Uncomment line below and debug to find bug.
-    // shoppingCart = getShoppingCart(events.toArray());
+
+    shoppingCart = getShoppingCart(eventStore, shoppingCartId);
 
     assertEquals(shoppingCartId, shoppingCart.id());
     assertEquals(clientId, shoppingCart.clientId());
-    assertEquals(ShoppingCartStatus.Confirmed, shoppingCart.status());
     assertEquals(2, shoppingCart.productItems().length);
+    assertEquals(ShoppingCart.Status.Confirmed, shoppingCart.status());
 
-    assertEquals(pricedPairOfShoes, shoppingCart.productItems()[0]);
-    assertEquals(pricedTShirt, shoppingCart.productItems()[1]);
+    assertEquals(shoesId, shoppingCart.productItems()[0].productId());
+    assertEquals(pairOfShoes.quantity(), shoppingCart.productItems()[0].quantity());
+    assertEquals(pricedPairOfShoes.unitPrice(), shoppingCart.productItems()[0].unitPrice());
+
+    assertEquals(tShirtId, shoppingCart.productItems()[1].productId());
+    assertEquals(tShirt.quantity(), shoppingCart.productItems()[1].quantity());
+    assertEquals(pricedTShirt.unitPrice(), shoppingCart.productItems()[1].unitPrice());
+
+    assertThat(shoppingCart.productItems()[0]).usingRecursiveComparison().isEqualTo(pricedPairOfShoes);
+    assertThat(shoppingCart.productItems()[1]).usingRecursiveComparison().isEqualTo(pricedTShirt);
+
+    var events = eventStore.readStream(ShoppingCartEvent.class, shoppingCartId);
+    assertThat(events).hasSize(5);
+    assertThat(events.get(0)).isInstanceOf(ShoppingCartOpened.class);
+    assertThat(events.get(1)).isInstanceOf(ProductItemAddedToShoppingCart.class);
+    assertThat(events.get(2)).isInstanceOf(ProductItemAddedToShoppingCart.class);
+    assertThat(events.get(3)).isInstanceOf(ProductItemRemovedFromShoppingCart.class);
+    assertThat(events.get(4)).isInstanceOf(ShoppingCartConfirmed.class);
   }
 }
