+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/solution2/ProductItems/ProductPriceCalculator.java
@@ -1,0 +1,8 @@
+package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems;
+
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems.ProductItems.PricedProductItem;
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems.ProductItems.ProductItem;
+
+public interface ProductPriceCalculator {
+  PricedProductItem calculate(ProductItem productItems);
+}
