+++ workshops/introduction-to-event-sourcing/solved/src/test/java/io/eventdriven/introductiontoeventsourcing/e05_business_logic/mutable/solution2/ShoppingCartEvent.java
@@ -1,0 +1,38 @@
+package io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2;
+
+import java.time.OffsetDateTime;
+import java.util.UUID;
+
+import static io.eventdriven.introductiontoeventsourcing.e05_business_logic.mutable.solution2.ProductItems.ProductItems.PricedProductItem;
+
+public sealed interface ShoppingCartEvent {
+  record ShoppingCartOpened(
+    UUID shoppingCartId,
+    UUID clientId
+  ) implements ShoppingCartEvent {
+  }
+
+  record ProductItemAddedToShoppingCart(
+    UUID shoppingCartId,
+    PricedProductItem productItem
+  ) implements ShoppingCartEvent {
+  }
+
+  record ProductItemRemovedFromShoppingCart(
+    UUID shoppingCartId,
+    PricedProductItem productItem
+  ) implements ShoppingCartEvent {
+  }
+
+  record ShoppingCartConfirmed(
+    UUID shoppingCartId,
+    OffsetDateTime confirmedAt
+  ) implements ShoppingCartEvent {
+  }
+
+  record ShoppingCartCanceled(
+    UUID shoppingCartId,
+    OffsetDateTime canceledAt
+  ) implements ShoppingCartEvent {
+  }
+}
